requirements.txt
altgraph==0.16.1
amqp==2.3.2
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
future==0.16.0
gevent==1.3.6
greenlet==0.4.15
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jsonfield==2.0.2
kombu==4.2.1
lxml==4.2.5
macholib==1.11
Markdown==3.0.1
mysqlclient==1.3.13
parsel==1.5.0
pefile==2018.8.8
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycurl==7.43.0.2
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyInstaller==3.4
pymongo==3.7.1
pyOpenSSL==18.0.0
pytz==2018.5
pywin32==224
pywin32-ctypes==0.2.0
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
Scrapy==1.5.1
scrapy-djangoitem==1.1.1
scrapyd==1.2.0
selenium==3.14.1
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
urllib3==1.23
vine==1.1.4
w3lib==1.19.0
zope.interface==4.5.0
